Fix spurious panic in `put_client_report`. (#1520)
This was another snapshot-isolation issue: in Repeatable Read, a write
can fail (due to a concurrent, conflicting write) but the conflicting
write can be unreadable. Generally, we want to force a retry when we
encounter such a situation.

(Discovered by analysis, I did not trigger this race condition. It would
require multiple concurrent uploads of the same-ID report to trigger.)