Move `client_timestamp_interval` from batch_aggregations to batches.
The incremental GC design expected batches to have a
`client_timestamp_interval` field; I tried writing the queries to read
from the sharded batch_aggregations table, but this was painful, so I
move it to batches instead and implement the design as written.
Eventually I want to merge `batches`, `batch_aggregations`, and
`outstanding_batches`; this is one step along that path.