Fix flaky argument parsing test (#1667)
The tests for `collect` binary argument handling were occasionally
failing. In one such run ([1]), we observe that the test randomly
generated an auth token `-7ekJePnK7WXktf92bK1CQ`, which starts with `-`
and confuses `clap`, which tries to treat it as a series of flags rather
than the value to `--dap-auth-token`. We rule this out by using the
`--dap-auth-token={token}` form, as these tests already did for task IDs
and other values that can start with `-`.

[1]: https://github.com/divviup/janus/actions/runs/5733188592/job/15537476867