Implement helper behavior
The aggregate init flow is where a task is advertised to the helper by the leader. The bulk of the
logic follows from there: we check if we're willing to opt-in to the task and then insert it into
the database before proceeding with the normal code flow. We use the peer aggregator's auth tokens
to authorize the request.

The aggregate continue flow is largely the same, we just make some sanity checks of the taskprov
header, use the peer aggregator's auth tokens, then proceed as normal.

The aggregate share flow requires that we propagate the peer aggregator fairly deep into its
function stack, so that we can use the peer aggregator's collector HPKE key to encrypt the aggregate
share. It otherwise works the same as the aggregate continue flow.

The task cache is a bit messy now--I will save refactoring it into a proper Cache in cache.rs for
a future PR to keep this one small.

Some stray .unwraps() from changing collector_hpke_config to nullable are here as well.