Backport taskprov-04 onto 0.subscriber-01. (#1742)
* taskprov: Implement helper behavior (#1675)

* aggregator_core changes for taskprov helper behavior

Major changes:
- New schema for taskprov peer aggregators. Stores all parameters that we must agree upon with our
  peer. Just has simple insertion and fetching. A future PR will be for a janus_cli wrapper so we
  can configure this. For auth tokens and collector HPKE config, we will use the peer aggregator
  values directly rather than copying them into a task.
- Make collector_hpke_config on tasks nullable, so that we can use the peer aggregator's collector
  HPKE config instead. The alternative is just inserting a junk key, but this makes task comparison
  in tests combersome.
- Introduce a wrapper taskprov::Task around the normal Task. This is so that we can insert
  taskprov-specific validation logic (i.e. empty auth tokens and collector hpke config are nullable).
- Fix `put_tasks` so that it appropriately returns MutationTargetExists error when there's a
  conflict. This is so that we can detect when multiple aggregator replicas try to insert the same
  task.
- VDAF verify key derivation logic is here as well.

* Miscellaneous changes

- Some helper/conversion functions.
- Add the taskprov DAP error type
- Add the taskprov header. It's in `core` since this would also be used by the client and collector.
- Add non_exhaustive to to Query (it should have been there in the first place).

* Implement helper behavior

The aggregate init flow is where a task is advertised to the helper by the leader. The bulk of the
logic follows from there: we check if we're willing to opt-in to the task and then insert it into
the database before proceeding with the normal code flow. We use the peer aggregator's auth tokens
to authorize the request.

The aggregate continue flow is largely the same, we just make some sanity checks of the taskprov
header, use the peer aggregator's auth tokens, then proceed as normal.

The aggregate share flow requires that we propagate the peer aggregator fairly deep into its
function stack, so that we can use the peer aggregator's collector HPKE key to encrypt the aggregate
share. It otherwise works the same as the aggregate continue flow.

The task cache is a bit messy now--I will save refactoring it into a proper Cache in cache.rs for
a future PR to keep this one small.

Some stray .unwraps() from changing collector_hpke_config to nullable are here as well.

* Taskprov tests

* Simple PR fixes

* Refactor common code between validation functions, reintroduce DB validation

* Simplify error field

* Give aggregator functions an already parsed task config

* Fix broken integration test.

The new validation logic introduce checks whether values are out of range for postgres. This code
inserts a task into Daphne, which doesn't have the same limitations. Just lower the number to fit
in i64.

* Accidentally broke this test

* Forgot to remove needless derivative

* Missing tracing on this function

* Dead test

* PR feedback

* Aggregator API endpoints for global HPKE key management (#1686)

* Aggregator API endpoints for global HPKE key management

* PR review feedback

* It would help to update the path in tests, too

* PR feedback, fully rename keypair -> config

* in_cluster: Create HPKE config in divviup-api first (#1693)

* Update in_cluster test w/ new HPKE config resource

* Bump version to 0.5.14

* cargo clippy

(*shakes fist at 1.65 clippy CI, yet again* -- this was another case of
a difference between current stable's clippy & 1.65's clippy.)

---------

Co-authored-by: Ameer Ghani <inahga@divviup.org>
Co-authored-by: David Cook <dcook@divviup.org>