Backport aggregator API changes and bug fixes into release/0.subscriber-01 (#1748)
* Split aggregator_api/src/lib.rs (#1683)

* Minor aggregator API cleanup (#1694)

* Deletion should return 204 on not found, not 404.

It's not really an error if we try to delete something and it's already been deleted. This also
makes the deletion request idempotent.

* Provide explicit return type in routes

* Remove needless Halt

I'm pretty sure this Halt isn't required and is just noisy. Other routes don't use it.

* Aggregator API error handling cleanup (#1698)

* Aggregator API error handling cleanup

As this crate grows, having to map each error is becoming burdensome.

Mainly turns `Error` into a enum from a struct. We can map errors into variants that handle HTTP
status codes for us, and use `#[from]` to let us use `?` on datastore errors.

We do lose a little bit of tracing in this PR. Datastore errors are unaffected, because datastore
methods are already traced. But handler-local internal error paths aren't well traced, some context
might be missing since we'll only see the final result. There's ways around this, but the internal
error cases look to be _very_ unlikely to happen (and likely should be unwraps anyways), so I don't
think it's worth spending much time on.

* PR feedback

* Fix flaky test that occasionally collides on random() (#1729)

* Add peer aggregator endpoints to aggregator API (#1689)

* Aggregator API endpoints for configuring taskprov peer aggregator

* Remove querystring as a dependency

* Align custom serde with other implementations

* Clippy

* Clippy lints that don't show up locally :shrug:

* Provide role and endpoint in request body

* Revert "Remove querystring as a dependency"

This reverts commit 74de8a752f7df8abc7b725ac959c7b05f1cfe22e.

* Remove serde_urlencoded

Cargo wants to remove random things :shrug:

* Remove unneeded error

* Fix panic when trying to get HPKE configs on a task that doesn't have any configured (#1738)

* Document taskprov extension and global HPKE keys (#1736)

* Document taskprov extension and global HPKE keys

* Typo

* Once you go taskprov, you never go back

* Spacing

Co-authored-by: Tim Geoghegan <timg@divviup.org>

* Typo

Co-authored-by: Tim Geoghegan <timg@divviup.org>

* Wording

Co-authored-by: Tim Geoghegan <timg@divviup.org>

* We're doing taskprov-04 for dap-02

---------

Co-authored-by: Tim Geoghegan <timg@divviup.org>

---------

Co-authored-by: Tim Geoghegan <timg@divviup.org>