DAP-05: Specialize to exactly two aggregators
This commit takes changes to various messages so that we handle exactly
one leader and exactly one helper, as specified in [DAP-05][1]. These
changes are voluminous but mostly mechanical.

This commit also adds variants for `Poplar1` to a couple of enums and
some macros, because later changes will need it.

[1]: https://datatracker.ietf.org/doc/draft-ietf-ppm-dap/05/

Part of #1669