Check collection job equivalence by query, not batch ID.
This allows current-batch requests to be reissued in fixed-size tasks,
simplifying implementation of a Collector.

While I'm at it, I also update `get_collection_job` to take a task ID in
addition to a collection job ID. This provides isolation between
different tasks, which is especially imporant now that the Collector
generates the collection job IDs -- otherwise, it would be trivial for
Collectors to generate the same collection job ID for two tasks, which
would cause some very strange behavior in Janus.