Release Janus `0.6.0-prerelease-1`
This release is being cut for integration testing, but tagged as a
pre-release because we anticipate making more disruptive changes
(especially SQL schema changes) before deploying it anywhere.

The Janus maintainers DO NOT recommend deploying this version as it will
be incompatible with eventual 0.6.x releases and no migration path will
be provided.