One VDAF verify key per task
The security analysis done for DAP-04 concluded that the VDAF verify key
must be agreed upon by aggregators before any reports may be aggregated
and cannot later be changed. That means a DAP task can only ever have a
single VDAF verify key. This commit simplifies the representation of
tasks in the various in-memory structures as well as the database to
reflect this. The biggest change is that we no longer have table
`task_vdaf_verify_keys` and instead add a row `vdaf_verify_key` to table
`tasks`.

Part of #1524, #1521