Change default auth token type to `BEARER` (#1962)
I don't think this makes a real difference, because all the code paths
that write auth tokens into the database should be setting the token
type explicitly anyway, and we switched the default type in the
aggregator API to `BEARER` a while ago. But this at least is consistent
with the other token tables and will be less confusing to readers.