Task rewrite: adopt `AggregatorTask` in taskprov
Adopts `janus_aggregator_core::task::AggregatorTask` in the `taskprov`
module and code that uses it in `janus_aggregator`.

Of particular interest is that we do away with
`janus_aggregator_core::taskprov::Task` and instead represent taskprov
tasks as a `janus_aggregator_core::task::Task` with
`AggregatorTaskParameters::TaskProvHelper`. Hopefully we'll be able to
further unify handling of taskprov and regular tasks in the future.

Part of #1524