only store collector auth token hash in leader
The leader only needs the hash of the collector auth token in order to
authenticate collection sub-protocol requests and thus we can avoid the
risk of storing the unhashed token.

This commit changes the `tasks` table to store a hash instead of the
token, and makes the corresponding change to
`janus_aggregator_core::task::AggregatorTask`. There are corresponding
changes to the aggregator API: the leader must now be provided the
collector auth token hash during task creation, and `TaskResp` now only
contains a collector auth token hash.

Additionally, we extend the aggregator API's config endpoint so that it
now can report a list of features that will be used by divvi-up API to
decide how to interact with it. The only such feature, which Janus
unconditionally advertises, is `token-hash`, indicating that Janus
stores token hashes instead of unhashed tokens where possible.

Part of #1509, #1524