Only store aggregator auth token hash in helper (#2062)
When acting as the helper, Janus only needs the aggregator auth token
hash to validate requests. It never makes aggregation sub-protocol
requests and thus doesn't need the actual token.

In this commit we change the schema of the `tasks` table so that helpers
may store an auth token hash instead of an auth token. These changes are
reflected in `janus_aggregator_core::task::AggregatorTask`. Finally we
make changes to the aggregator API to accommodate this: the helper
reports the aggregator auth token it generates during task provisioning,
but doesn't store that token and cannot subsequently provide it when
responding to `GET /tasks/:task_id` requests. Fortunately that field in
`TaskResp` was already optional, and divviup-api already ignored it
except in the creation case.

Part of #1509, 1524