database: record who modified rows and when
Adds rows `updated_by` and `created_at` to all database tables.
`updated_by` is set to the transaction name passed to
`Datastore::run_tx_with_name`. `updated_at` and `created_at` are set to
the time at which the row is mutated. `updated_at` is only present for
tables whose rows are mutable.

These values aren't used for anything in Janus and so do not appear in
any Janus data structure. They are intended to aid operators debugging
or otherwise examining activity in Janus.

Resolves #1858