adopt new `divviup-client`
The PRs related to https://github.com/divviup/divviup-api/issues/542
implement the new concept of "collector credential". This entails
changes to the API for provisioning tasks via divviup-api that we use in
e2e integration tests. In this commit we adopt the new `divviup-client`
and fix up its usage.

Part of #2071