release/0.5: adopt new `divviup-api` API (#2151)
The PRs related to https://github.com/divviup/divviup-api/issues/542
implement the new concept of "collector credential". This entails
changes to the API for provisioning tasks via divviup-api that we use in
e2e integration tests. On `release/0.5`, we don't use the
`divviup-client` crate, so we update the client baked into Janus.

Another curiosity is that while `divviup-api` will mint a collector auth
token in response to our request to create a collector auth token, that
auth token won't be used and instead the one generated by the Janus
aggregator API gets plumbed back to us in the task creation request.

Part of #2071