Fix repeated collection for time-interval. (#2222)
This was caused by the batch_aggregations soft-delete condition not
respecting batches.batch_interval, which is used in time-interval tasks
to allow coherent GC semantics even for batches including 0 reports
(which time-interval must allow). This bug could specifically be
triggered only if a collection request was made for a time interval
including a minimal time interval with 0 reports, and then repeated. In
that case, the empty batch aggregations would be skipped due to the GC
condition, which would then cause us to try writing them again, leading
to a MutationTargetAlreadyExists error which would fail the collection
job step until the collection job was abandoned.

This bug could only occur in time-interval: fixed-size tasks will always
have more than 0 reports per batch.

I also augmented our integration tests to try collecting the same batch
twice, to ensure we do not regress on this behavior.