Integration test for rate limits
Adds tests to exercise Divvi Up's rate limiting to the existing
`in_cluster` test module. The test setup reads in rate limit
configuration from a file whose path is specified via environment
variable, akin to the existing `JANUS_E2E_*` variables.

These tests don't actually exercise any part of Janus itself, but they
rely on the test support for forwarding ports and otherwise working with
a pair of Januses in a Kubernetes cluster, so this is a convenient place
to define them.