GC: add config to group GC in tx's, control concurrent tx's.
Two new settings are introduced to the GC config:
 * tasks_per_tx: controls how many tasks are grouped into the same
   transaction for purposes of garbage collection. If left unset, each
   task's GC work is run in its own transaction.
 * concurrent_tx_limit: controls how many concurrent database
   transactions are allowed for GC. If left unset, there is no limit to
   the number of concurrent transactions used for GC.

If neither of these new options are configured, the behavior will be
left as before -- each task will be GC'ed in its own transaction, with
no limit to the number of concurrent transactions.

Note that grouping GC for different tasks into a single transaction
risks failures (or retries) in the GC process for one task impacting GC
for other, independent tasks.