Run E2E integration tests against cloud env (#2382)
Extends the test harness in
`janus_integration_tests::integration::in_cluster` so that it can run
tests against remote instances of divviup-api, a Janus leader and a
Janus helper.

Concretely, this means connecting to those services at some remote URL
rather than over a `kubectl port-forward` to a local Kind cluster. Past
that, we also exercise less of the `divviup-api` surface.

In Kind, we do everything from creating an account through provisioning
a task. This works because the `divviup-api` used in that context has
authentication disabled and also because that entire environment is
understood to be ephemeral -- we can litter it with tasks and paired
aggregators and anything else we want and not clean up, because our
expectation is that the entire Kubernetes cluster and all its storage
will be destroyed.

But in this new configuration, we expect to run the test in long-running
staging environments, or even production. While I'm not concerned about
adding new tasks, reports and aggregatons there -- task expiry and
garbage collection will eventually clean those up -- I don't want to
create new accounts and paired aggregators on every test run. That means
that the test needs to be run with a handful of `JANUS_E2E_` environment
variables defined that provide an authentication token and IDs for
various resources to be used. See the doccomment on
`InClusterJanusPair::new_in_cloud` for details.

This change also adds yet more code to Janus that is very, very useful
for testing Divvi Up, but not necessarily relevant to Janus (after all,
someone could be running Janus without `divviup-api` or without
Kubernetes). I'm interested in moving this chunk of test harness
elsewhere, but for now this is the best way to get a reviewable diff.

Relevant to https://github.com/divviup/janus-ops/issues/86