Fix writing aggregation jobs touching GC'ed batches.
This issue should only exist in the time-interval query type, as
fixed-size is arranged such that aggregation jobs touching a given batch
must be GC'ed before the batch. I include a guard to ensure that the new
codepath is only taken in the expected case of an already-GC'ed batch
for a time-interval query, as otherwise we might drop batch writes if we
fell into it unexpectedly.