Fixed-size: return each batch to at most one current-batch query. (#2391)
Part of DAP-09 implementation (#2389). With this change, Collectors are
expected to remember the collection job IDs they generate before sending
the collection request. If the Collector encounters a transient failure,
it can then resend the request (with the same collection job ID), and
rely on request idempotency for things to work out.

The upside of this change is that Collectors no longer need to have
logic to deduplicate multiple current-batch requests that are mapped to
the same underlying batch, and they can generate current-batch requests
at higher rates without risking more duplicated collections.