Datastore: store report info with report aggregation.
This will be used to allow aggregation jobs to carry along the data they
use for processing, which will insulate them against potential changes
in light of e.g. client-report GC. Specifically, this will be used to
ensure that retries of aggregation job steps will lead to the same
request being made, even if some of the relevant client reports have
been GC'ed in the meantime.