leader deletes abandoned aggregation job in helper
Per DAP-09, when the leader gives up on an aggregation job, it SHOULD
delete that job in the helper. Our leader (specifically
`aggregation_job_driver`) now does this, and we wire up a simple handler
for deleting aggregation jobs on the helper side. This handler does
nothing but mark the relevant row in the aggregations job as deleted.
This could eventually be used by Janus garbage collection to more
aggressively evict data before its expiry.

Part of #2389