taskprov-06: update messages.
Many messages are now encoded as opaque values, to allow the message to
parse even if we don't know about all of the query types/VDAFs/DP
mechanisms in play. TBH, I'm not sure this is required -- in all cases,
the relevant messages include a discriminator byte, which could be used
to fail parsing before the parser can become confused.