[0.6] Aggregation: repeat same request on retries. (#2556)
This is implemented by copying the necessary report information into the report aggregations at time of aggregation job creation. This means that client-report GC will not change the request that is made.

Over & above 0.7 version of this functionality, the 0.6 version retains the ability to work with aggregation jobs which include client report data in the `client_reports` table.