Update retry_http_request to retry reading response body.
Before, the response body was not read by `retry_http_request`; instead,
the caller was expected to receive the `reqwest::Response` and read the
body themselves. At all call sites, failures in reading the body would
not be retried, meaning we only _mostly_ had a retry strategy.

Also, update the semantics of `retry_http_request` to match the
documented semantics:
 * On success, return the response (now represented as an
   `HttpResponse`).
 * On an HTTP-level error, return an `HttpErrorResponse`.
 * On a network-level error, return the `reqwest::Error` directly.