Remove batches table.
The goal of this change is to reduce write contention, particularly on
creation or stepping of aggregation jobs. Previously, every aggregation
job creation or finalization required updating the relevant (unsharded)
batches row. This was a source of significant write contention.

All uses of the unsharded batches table have been migrated to the
sharded batch aggregations table. In particular:
 * Batches updates, particularly the state value, were used to avoid the
   possibility of aggregating into a collected batch even in the face of
   concurrent aggregation/collection. Now, the batch aggregations' state
   value is used.
 * Batches included an outstanding_aggregation_jobs value which was used
   to track how many aggregation jobs still needed to be completed. This
   is replaced by sharded aggregation_jobs_{created,terminated} counters
   in batch aggregations.
 * The client_timestamp_interval contained in the batch was used for GC.
   Now, a sharded client timestamp interval is included in batch
   aggregations, and is used for this purpose.

Note that, due to the particulars of Postgres' concurrency model, we now
have to poll collection jobs for readiness in the collection job driver;
I am not aware of a method of implementing a sharded counter where we
can be guaranteed that at least one transaction updating the counter
will realize when the counter has reached a certain value.

I also fixed a few bugs and cleaned up some technical debt along the
way. In particular:
 * As Leader & Helper aggregation job writes are now identical, they
   both use aggregation_job_writer. The accumulator previously used for
   Helper writes has been removed.
 * Helper aggregation init could accumulate shares, then fail the report
   (e.g. on replay), then still write the accumulated value including
   the failed report's contribution. This was a privacy bug.
 * The check for all-batches-complete for a given collection job was
   racy. This was a correctness bug.
 * Abandoning an aggregation job could mark report aggregations as
   BatchCollected if they fell into a collected batch. Now they are left
   untouched, as intended, to allow investigation of abandoned
   aggregation jobs.
 * Collection job interval computation is moved to collection job
   stepping; this removes the need for collection job retrieval to read
   the batch aggregations, and imposes no additional reads on the
   collection job stepping code as it was already reading the batch
   aggregations.
 * Several tables had an unused "trace_context" column. This column was
   never read or written, so I removed it.