Datastore: treat writes to logically GC'ed rows as new writes. (#2818)
Previously, such writes would lead to a MutationTargetAlreadyExists
error. However, the same error could happen if another transaction
concurrently wrote the same entity, which would have to be manually
retried (by `tx.retry()`) by application code if it was sure the row
shouldn't exist. This could lead to infinitely-repeated retries if the
MutationTargetAlreadyExists error was due to the row being GC'ed.

Now, we perform an upsert if the existing row is logically GC'ed. This
allows overwriting logically-nonexistent rows with a put, but more
importantly, this should allow concurrent writes to receive a
serialization failure as expected, rather than a
MutationTargetAlreadyExists error. This removes the need for
application-level retries entirely, which should remove the possibility
of further infinite-retry bugs.

Also:
* Remove delete-after-put for all put operations. We will wait for the
  GC run to delete these rows. This is a tradeoff in complexity for the
  new upserts.
* put_report_share renamed to put_scrubbed_report, and give it the same
  semantics as all other put_foo datastore methods. Note that this
  required some changes to the application code using this
  functionality, please review carefully.
* put_report_aggregation & similar now properly reports
  MutationTargetAlreadyExists.
* put_client_report: fix the expiry check.
* get_report_aggregation renamed to
  get_report_aggregation_for_report_id, now test-only. This was already
  only used for tests.
* Simplify a few GC conditions: several GC conditions had an unnecessary `COALESCE(..., FALSE)` around
  the GC check. This condition was unnecessary because a WHERE clause
  filters rows whose condition evaluates to either FALSE or NULL.