Feature: signed HPKE configs. (#2880)
This PR adds the ability to sign the responses from the /hpke_config
endpoints. Currently, only ECDSA P-256 signatures are supported. The
resulting signature is added to a "x-hpke-config-signature" response
header, as an unpadded URL-safe base64 encoding of the signature bytes.

This new feature is controlled via configuration. By default, the HPKE
configs are not signed. Specifying the new hpke_config_signing_key
configuration option in the aggregator config will cause the HPKE
configs to be signed. This configuration option expects a PEM-encoding
of a DER-encoding of an RFC 5915 ECPrivateKey structure, containing a
P-256 key. This format was chosen for compatibility, as it is commonly
generated by openssl.