Pool wait time metric (#2953)
Create a new pool wait metric that captures only the time spent acquiring a
connection from the pool. This should signal if the connection pool size is
too small.

Also modifies the current transaction duration metric to avoid counting the
time spent waiting on the pool. Now, if this metric is out of whack, it should
only signal that the query(ies) are inefficient or the database is overloaded.