Collection job driver: reduce work required to detect step readiness. (#2957)
Specifically, we now check if there are unaggregated reports, or
unfinished aggregation jobs, in the same pipelined round-trip as reading
the task and collection job. The outstanding aggregation job count is
determined without reading the set of batch aggregations, which can be
quite large in some cases as they include partial aggregate shares.

We also now check for a matching, already-completed collection job
before reading the batch aggregations.