Task aggregator cache: remove redundant reads on concurrent requests.
Previously, if multiple requestors concurrently retrieved the same task
from the task aggregator cache, we would query the DB for the task for
each requestor.

Now, we use an additional layer of per-task locking so that, in the
concurrent-requestor scenario, one requestor will read the task from the
DB, and the remaining tasks will wait on that requestor's value.