Helper: parallelize aggregation computations.
Previously, we used a rayon threadpool for aggregation computations, but
the computations within the handling of a single aggregation job were
still serialized. Now, we use a parallel iterator for handling each
report share in turn, allowing the VDAF evaluations to be performed in
parallel.

Also, remove the (unstated!) requirement in aggregation_job_writer than
report aggregations be provided in order. This eases parallelization of
the aggregation job continuation logic.