Implement cancellation on disconnect.
This will be very helpful when we have requests which timeout, which is
presented to Janus as a client disconnect. Otherwise, we would continue
processing the request until it is complete.

Update Helper aggregation methods, which use rayon to parallelize
processing, to respect cancellation. Specifically, the `receiver` will
be dropped, causing the producer's `sender` to return a SendError, which
will cause `try_for_each_with` to stop processing.