Log agg job init forbidden mutations.
When we detect that an aggregation job ID is being re-used, we check
that the incoming request matches whatever request we previously saw,
and reject it with `Error::ForbiddenMutation`, which is rendered as HTTP
409 Conflict.

We now record an event when this occurs for aggregation job
initialization. The event contains the expected request hash, the
incoming request's hash and the list of `ReportMetadata`s for both.

The event is logged in a new span `aggregation_job_init_forbidden_mutation_span`
at the DEBUG level. This event will be logged if a suitable filter is
set up. For example, running Janus with `RUST_LOG=debug` or a `RUST_LOG`
environment variable containing this string will do the trick:

`janus_aggregator::aggregator[aggregation_job_init_forbidden_mutation_span]=debug`

Logging the forbidden mutation is expensive because it requires fetching
all the `report_aggregations` rows for the existing aggregation job, so
this is gated on the `log_forbidden_mutations` option.