Log agg job init forbidden mutations. (#3129)
When we detect that an aggregation job ID is being re-used, we check that the incoming request matches whatever request we previously saw, and reject it with Error::ForbiddenMutation, which is rendered as HTTP 409 Conflict.

We now record an event when this occurs for aggregation job initialization. The event contains the expected request hash, the incoming request's hash and the list of ReportMetadatas for both.

Because the resulting event object is quite large, it gets written to a file, created under the directory specified in the log_forbidden_mutations option.

Logging the forbidden mutation is expensive because it requires fetching all the report_aggregations rows for the existing aggregation job, so this is gated on the log_forbidden_mutations option.

Note that there are also cases in aggregation job continuation and collection job creation where we check for resource mutation and return 409 Conflict. This commit only logs on the aggregation job init path because that's what we're trying to debug right now.