Fix Helper aggregation-initialization report-replayed check.
Specifically, we now check only if the corresponding row exists in the
`client_reports` table, rather than ignoring that fact and looking for a
corresponding `report_aggregations` row. This means we no longer respect
whether the other aggregation used a different aggregation parameter,
but that's fine -- Janus now supports only a single aggregation per
report.

This fixes a few bugs:
* If the same report was repeatedly submitted in different aggregation
  jobs, the first instance would be aggregated; the second instance
  would be marked as a replay; and the third and further instances would
  result in an HTTP 500 error, as
  `check_other_report_aggregation_exists` mistakenly expected at most a
  single matching `report_aggregations` row to exist.
* If the same report was submitted in more than one aggregation job
  concurrently, Janus would have aggregated both. This was a privacy
  bug.