Improve efficiency of collection-artifact GC.
Previously, we would need to scan the batch_aggregations table to
fulfill batches_to_delete, which was especially painful if there were
not enough batches to delete because we would have to scan the entire
table to determine that.

Now, we do a quick first pass to determine candidate batches to delete,
with the aid of a new index.