Fix test failure.
I'm not sure why adding an index caused problematic behavior to begin to
occur, but the basic problem was the lack of sort on the result of
`get_batch_aggregations_for_collection_identifier`; Postgres could
always have returned these batch aggregations in any order, and with the
addition of the index contained in this PR, it did in fact start
reordering results with respect to `ord`.

Other than the tests passing, I don't imagine this will be a problem in
a deployed environment, as all uses of the combined batch aggregations
for a given batch merge them in an order-irrelevant way.