Helper agg init: Remove copy of report share data.
Previously, we copied report share data into the transaction callback
(re-copying it on transaction retry). Notably, this required us to keep
two copies of the output shares in memory. This was done to allow
modification of the report aggregations based on information retrieved
from the datastore, i.e. if the report was replayed.

Now, we remove the copy by passing an Arc of the relevant data into the
transaction callback. Modification of the report aggregation is
implemented via a Cow, which will be borrowed from the data in the Arc
in the common case of no modification. The aggregation job writer is
augmented with the ability to accept report aggregations already stored
in a Cow, and is taught to be smart enough not to double-wrap Cow around
its already-Cow'ed internal state.